Actually, treat mention in reply as mention type instead
This is what the tumblr blog settings do and i think it is more expected. At first i put it under reply because it is a reply to you but that would be a reply to your own post. Mentions can be in reblogs posts and replies and as a result this goes here i think.